fix: norm
